+++ src/app/(main)/(owner)/[id]/device/[deviceId]/page.tsx
@@ -1,22 +1,5 @@
-"use client";
-
-import { useStoreContext } from "@/providers/storeProvider";
-import { useRouter, useParams } from "next/navigation";
-import { useEffect } from "react";
+import DevicePage from "../_components/_template/DevicePage";
 
 export default function Page() {
-  const navigate = useRouter();
-  const params = useParams();
-
-  const deviceId = params?.deviceId;
-
-  const { storeId } = useStoreContext();
-
-  useEffect(() => {
-    if (deviceId) {
-      navigate.replace(`/${storeId}/device/${deviceId}`);
-    }
-  }, [deviceId, navigate, storeId]);
-
-  return null;
+  return <DevicePage />;
 }
